move context-related functions from editor_helpers.ts to Context
